Добавил прелоудеры на кнопки

+++ src/containers/Home/Home.tsx
@@ -8,40 +8,55 @@ import MealItem from '../../components/MealItem/MealItem.tsx';
 const Home = () => {
   const [meals, setMeals] = useState<IMeal[]>([]);
   const [total, setTotal] = useState<number>(0);
+  const [loading, setLoading] = useState<boolean>(false);
 
   const fetchMeals = useCallback(async () => {
-    const response:{data: IMealAPI} = await axiosAPI<IMealAPI>('meal.json');
-    const mealList = response.data;
+    try {
+      setLoading(true);
+      const response:{data: IMealAPI} = await axiosAPI<IMealAPI>('meal.json');
+      const mealList = response.data;
 
-    if (mealList === null) {
-      setMeals([]);
-      return;
+      if (mealList === null) {
+        setMeals([]);
+        setTotal(0);
+        return;
+      }
+
+      if (mealList) {
+        const mealFromAPI = Object.keys(mealList).map((mealKey) => {
+          return {
+            ...mealList[mealKey],
+            id: mealKey,
+          };
+        });
+
+        const totalCalories = mealFromAPI.reduce((acc, meal) => {
+          acc += Number(meal.kcal) | 0;
+          return acc;
+        },0);
+        setTotal(totalCalories);
+        setMeals(mealFromAPI);
+      }
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setLoading(false);
     }
 
-    if (mealList) {
-      const mealFromAPI = Object.keys(mealList).map((mealKey) => {
-        return {
-          ...mealList[mealKey],
-          id: mealKey,
-        };
-      });
-
-      const totalCalories = mealFromAPI.reduce((acc, meal) => {
-        acc += meal.kcal | 0;
-        return acc;
-      },0);
-      setTotal(totalCalories);
-      setMeals(mealFromAPI);
-    }
   },[]);
 
 
   const deleteMeal = useCallback (async (id: string) => {
     try {
-      await axiosAPI.delete(`meal/${id}.json`);
-      await fetchMeals();
+      if (window.confirm('Are you sure you want to delete meal?')) {
+        setLoading(true);
+        await axiosAPI.delete(`meal/${id}.json`);
+        await fetchMeals();
+      }
     } catch (error) {
       console.error(error);
+    } finally {
+      setLoading(false);
     }
   },[fetchMeals]);
 
@@ -50,19 +65,26 @@ const Home = () => {
   },[fetchMeals]);
 
 
-
   return (
     <>
-      <Grid container spacing={2}>
-        {meals.map((meal) => (
-          <MealItem
-            key={meal.id}
-            meal={meal}
-            onDeleteMeal={deleteMeal}
-          />
-        ))}
-        <div>Количество калорий: {total}</div>
-      </Grid>
+        <>
+          <Grid container spacing={2}>
+            {meals.map((meal) => (
+              <MealItem
+                key={meal.id}
+                meal={meal}
+                onDeleteMeal={deleteMeal}
+                deleteLoading={loading}
+              />
+            ))}
+            <div>{total ? (
+              <>
+                Количество калорий: {total}
+              </>
+            ) : null}
+            </div>
+          </Grid>
+        </>
     </>
   );
 };
